requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
django-cors-headers==2.4.1
django-crispy-forms==1.7.2
django-filter==2.1.0
django-markdown-deux==1.0.5
django-rest-multiple-models==2.1.3
django-tinymce==2.8.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
idna==2.8
Markdown==3.0.1
markdown2==2.3.7
Pillow==5.4.1
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
